RG-2542 Split impl into multiple files to improve the TS performance
