+++ src/table/table-const.ts
@@ -1,0 +1,12 @@
+import {createContext} from 'react';
+
+import type {TableProps} from './table';
+
+export const TablePropsContext = createContext<TableProps<unknown> | null>(null);
+export const ColumnIndexContext = createContext<number>(-1);
+export const CollapseItemIntoSpacerContext = createContext<(height: number) => void>(() => {});
+
+export const defaultRowHeight = 37;
+export const defaultLookaheadPx = 400;
+export const defaultRetentionMarginPx = 450;
+export const defaultMinScrollAndResizeDeltaPx = 50;
