+++ src/table/table.tsx
@@ -1,27 +1,9 @@
-/* eslint-disable max-lines */
-import React, {
-  type Context,
-  createContext,
-  useContext,
-  type HTMLAttributes,
-  Fragment,
-  type ReactNode,
-  useRef,
-  useEffect,
-} from 'react';
-import classNames from 'classnames';
-import unsortedIcon from '@jetbrains/icons/unsorted-12px';
-import arrowDownIcon from '@jetbrains/icons/arrow-12px-down';
-import arrowUpIcon from '@jetbrains/icons/arrow-12px-up';
-import trashIcon from '@jetbrains/icons/trash-12px';
+import Table from './table-component';
 
-import Icon from '../icon/icon';
-import {IntersectionObserverContext, useIsIntersectingListener} from '../global/intersection-observer-context';
-import {SpacerRow, useTableVirtualize} from './table-virtualize';
-
+import type {ReactNode} from 'react';
 import type Selection from '../legacy-table/selection';
 
-import styles from './table.css';
+export default Table;
 
 export interface TableProps<T> {
   /**
@@ -308,341 +290,4 @@ export interface Column<T> {
 export interface DefaultRowRendererProps<T> {
   item: T;
   index: number;
-}
-
-export const TablePropsContext = createContext<TableProps<unknown> | null>(null);
-
-export const ColumnIndexContext = createContext<number>(-1);
-
-export const CollapseItemIntoSpacerContext = createContext<(height: number) => void>(() => {});
-
-const defaultRowHeight = 37;
-const defaultLookaheadPx = 400;
-const defaultRetentionMarginPx = 450;
-const defaultMinScrollAndResizeDeltaPx = 50;
-
-/**
- * The new Table component. Use it instead of tables in the `legacy-table` folder.
- *
- * Minimal usage requires the following props:
- * - `data`
- * - `getKey`
- * - `columns`
- *   - `key`
- *   - `renderCell` (not required, but usually needed)
- *
- * ## Selection
- *
- * Following three props support the selection:
- *
- * - `selection`
- * - `isClickable`
- * - `onItemClick`
- *
- * Only `selection` is required: you can display and modify selection your way, e.g., via
- * checkboxes in cells.
- *
- * ## Sorting
- * You need the following to support sorting:
- *
- * - Include `<SortButton />` in a column header
- * - Set initial `Column.sortOrder` to `none`, `ascending`.
- * Do not leave `undefined` for the accessibility reasons.
- * - Handle `TableProps.onSort` callback in the client code. It is expected
- * to update `columns`, by setting the new `sortOrder` value for
- * the corresponding column, and updating the data accordingly.
- *
- * ## Deleting columns
- * You need the following to support deleting columns:
- *
- * - Make sure the `column` has a proper `name` or `key` prop, which will be
- * automatically included in the aria-label of `<DeleteColumnButton />`.
- * - Include `<DeleteColumnButton />` in a column header
- * - Handle `TableProps.onColumnDelete` callback in the client code. It is expected
- * to update `columns` by removing the corresponding column.
- *
- * ## Row virtualization
- *
- * To render only rows near the viewport and replace others with spacers, use:
- *
- * - `virtualizeRows`
- * - `scrollerRef` — required when the scrollable container is not the whole document
- * - `estimateHeight` — recommended when rows are expected to be taller than
- *   the default height (e.g. multiline or custom content)
- * - Fine-tuning props: `lookaheadPx`, `retentionMarginPx`, `minScrollAndResizeDeltaPx`
- */
-export default function Table<T>(props: TableProps<T> & HTMLAttributes<HTMLTableElement>) {
-  const {
-    data,
-    columns,
-    renderItem,
-    virtualizeRows = false,
-    scrollerRef,
-    estimateHeight = () => defaultRowHeight,
-    lookaheadPx = defaultLookaheadPx,
-    retentionMarginPx = defaultRetentionMarginPx,
-    minScrollAndResizeDeltaPx = defaultMinScrollAndResizeDeltaPx,
-    className,
-    theadClassName,
-    theadTrClassName,
-    tbodyClassName,
-  } = props;
-
-  const tableRef = useRef<HTMLTableElement | null>(null);
-
-  const {virtualItems, intersectionObserverHandle, collapseItemIntoSpacer} = useTableVirtualize({
-    enabled: virtualizeRows,
-    length: data.length,
-    scrollerRef,
-    tableRef,
-    estimateHeight,
-    lookaheadPx,
-    retentionMarginPx,
-    minScrollAndResizeDeltaPx,
-  });
-
-  return (
-    <TablePropsContext.Provider value={props as TableProps<unknown>}>
-      <IntersectionObserverContext.Provider value={intersectionObserverHandle}>
-        <table className={classNames(styles.table, className)} ref={tableRef}>
-          <thead className={theadClassName}>
-            <tr className={classNames(styles.headerRow, theadTrClassName)}>
-              {columns.map((column, columnIndex) => (
-                <th
-                  key={column.key}
-                  className={classNames(styles.headerCell, column.thClassName)}
-                  aria-sort={column.sortOrder}
-                >
-                  <ColumnIndexContext.Provider value={columnIndex}>
-                    {column.renderHeader?.() ?? column.name ?? String(column.key)}
-                  </ColumnIndexContext.Provider>
-                </th>
-              ))}
-            </tr>
-          </thead>
-
-          <tbody className={tbodyClassName}>
-            {virtualItems.map(virtualItem => {
-              if (virtualItem.type === 'spacer') {
-                return <SpacerRow key={virtualItem.key} spacer={virtualItem} colSpan={columns.length} />;
-              }
-
-              const index = virtualItem.index;
-              const item = data[index];
-              const key = props.getKey(item, index);
-              return (
-                <CollapseItemIntoSpacerContext.Provider
-                  value={height => collapseItemIntoSpacer(index, height)}
-                  key={key}
-                >
-                  {renderItem ? (
-                    <Fragment>{renderItem(item, index)}</Fragment>
-                  ) : (
-                    <DefaultRowRenderer item={item} index={index} />
-                  )}
-                </CollapseItemIntoSpacerContext.Provider>
-              );
-            })}
-          </tbody>
-        </table>
-      </IntersectionObserverContext.Provider>
-    </TablePropsContext.Provider>
-  );
-}
-
-const INDENT_SIZE = 16;
-
-/**
- * The default row renderer used when `renderItem` is not provided.
- * You can also use it as a fallback in a custom `renderItem` implementation.
- */
-export function DefaultRowRenderer<T>({item, index}: DefaultRowRendererProps<T>) {
-  const tableProps = useContext(TablePropsContext as Context<TableProps<T>>);
-
-  const rowRef = useRef<HTMLTableRowElement | null>(null);
-
-  const collapseItemIntoSpacer = useContext(CollapseItemIntoSpacerContext);
-  useIsIntersectingListener(rowRef, isIntersecting => {
-    if (rowRef.current && !isIntersecting) {
-      collapseItemIntoSpacer(rowRef.current!.getBoundingClientRect().height);
-    }
-  });
-
-  const {columns, selection, isClickable, onItemClick, getItemLevel, tbodyTrClassName} = tableProps;
-  const clickable = isClickable?.(item, index);
-  const selected = selection?.isSelected(item);
-  const level = getItemLevel?.(item, index) ?? 0;
-
-  function getItemDependentClassName(className: undefined | string | ((item: T, index: number) => string)) {
-    if (!className) {
-      return undefined;
-    }
-
-    return typeof className === 'function' ? className(item, index) : className;
-  }
-
-  function onPointerUp(e: React.PointerEvent<HTMLTableRowElement>) {
-    if (!onItemClick) return;
-
-    const {target} = e;
-    const isActiveElement = target instanceof Element && target.closest('a, button, input, select');
-    if (isActiveElement) return;
-
-    onItemClick(e.nativeEvent, item, index);
-  }
-
-  function onKeyDown(e: React.KeyboardEvent<HTMLTableRowElement>) {
-    if (e.key === 'ArrowUp' || e.key === 'ArrowDown') {
-      const step = e.key === 'ArrowUp' ? -1 : 1;
-      // eslint-disable-next-line yoda
-      for (let i = index + step; 0 <= i && i < tableProps.data.length; i += step) {
-        if (tableProps.isItemFocusable?.(tableProps.data[i], i)) {
-          tableProps.onItemFocus?.(tableProps.data[i], i);
-          break;
-        }
-      }
-    }
-  }
-
-  const focused = selection?.isFocused(item);
-
-  useEffect(() => {
-    if (focused) rowRef.current?.focus();
-  }, [focused]);
-
-  function onBlur() {
-    if (focused) tableProps.onItemFocus?.(null, -1);
-  }
-
-  return (
-    <TableRow
-      ref={rowRef}
-      className={classNames(
-        getItemDependentClassName(tbodyTrClassName),
-        clickable && styles.clickableRow,
-        selected && styles.selectedRow,
-      )}
-      onPointerUp={onPointerUp}
-      onKeyDown={onKeyDown}
-      tabIndex={focused ? 0 : undefined}
-      onBlur={focused ? onBlur : undefined}
-    >
-      {columns.map((column, columnIndex) => (
-        <TableCell
-          key={column.key}
-          className={getItemDependentClassName(column.tdClassName)}
-          style={column.indent && level > 0 ? {paddingInlineStart: `${level * INDENT_SIZE}px`} : undefined}
-        >
-          {column.renderCell?.(item, index) ?? getDefaultCellValue(item, columnIndex)}
-        </TableCell>
-      ))}
-    </TableRow>
-  );
-}
-
-function getDefaultCellValue<T>(item: T, columnIndex: number) {
-  if (Array.isArray(item)) {
-    return String(item[columnIndex] ?? '');
-  }
-
-  if (item !== null && typeof item === 'object') {
-    return String(Object.values(item)[columnIndex] ?? '');
-  }
-
-  if (columnIndex === 0) {
-    return String(item);
-  }
-
-  return '';
-}
-
-/**
- * A helper `<tr>` component for custom `renderItem` implementations.
- * Applies standard row classnames, but not data-dependent `tbodyTrClassName`.
- */
-export function TableRow(props: {ref?: React.Ref<HTMLTableRowElement>} & HTMLAttributes<HTMLTableRowElement>) {
-  const {ref, className, ...restProps} = props;
-  const classes = classNames(styles.row, className);
-  return <tr ref={ref} className={classes} {...restProps} />;
-}
-
-/**
- * A helper `<td>` component for custom `renderItem` implementations.
- * Applies standard cell classnames, but not data-dependent `tdClassName`.
- */
-export function TableCell(props: HTMLAttributes<HTMLTableCellElement>) {
-  const {className, ...restProps} = props;
-  const classes = classNames(styles.cell, className);
-  return <td className={classes} {...restProps} />;
-}
-
-/**
- * Include it in a column header to make the column sortable.
- * Handle clicks with `TableProps.onSort`.
- */
-export function SortButton<T>(props: HTMLAttributes<HTMLButtonElement>) {
-  const tableProps = useContext(TablePropsContext as Context<TableProps<T>>);
-  const columnIndex = useContext(ColumnIndexContext);
-  const column = tableProps?.columns[columnIndex];
-  if (!tableProps || !column) {
-    return null;
-  }
-
-  const sortOrder = column.sortOrder ?? 'none';
-  // eslint-disable-next-line no-nested-ternary, prettier/prettier
-  const glyph = sortOrder === 'none' ? unsortedIcon
-    : sortOrder === 'ascending' ? arrowUpIcon
-    : arrowDownIcon;
-
-  const {className, children, onClick, ...restProps} = props;
-
-  function handleClick(e: React.MouseEvent<HTMLButtonElement>) {
-    onClick?.(e);
-    if (!e.defaultPrevented) {
-      const sequence = ['none', 'ascending', 'descending'] satisfies SortOrder[];
-      const nextOrder = sequence[(sequence.indexOf(sortOrder) + 1) % sequence.length];
-      tableProps.onSort?.(columnIndex, nextOrder);
-    }
-  }
-
-  return (
-    <button type='button' className={classNames(styles.headerButton, className)} onClick={handleClick} {...restProps}>
-      {children} <Icon glyph={glyph} aria-hidden />
-    </button>
-  );
-}
-
-/**
- * Include it in a column header to make the column deletable.
- * Beware that `column.name ?? String(column.key)` is used in the aria-label.
- * Handle clicks with `TableProps.onColumnDelete`.
- */
-export function DeleteColumnButton<T>(props: HTMLAttributes<HTMLButtonElement>) {
-  const tableProps = useContext(TablePropsContext as Context<TableProps<T>>);
-  const columnIndex = useContext(ColumnIndexContext);
-  const column = tableProps?.columns[columnIndex];
-  if (!tableProps || !column) {
-    return null;
-  }
-
-  const {className, onClick, ...restProps} = props;
-
-  function handleClick(e: React.MouseEvent<HTMLButtonElement>) {
-    onClick?.(e);
-    if (!e.defaultPrevented) {
-      tableProps.onColumnDelete?.(columnIndex);
-    }
-  }
-
-  return (
-    <button
-      type='button'
-      className={classNames(styles.headerButton, styles.deleteColumnButton, className)}
-      onClick={handleClick}
-      aria-label={`Delete column ${column.name ?? String(column.key)}`}
-      {...restProps}
-    >
-      <Icon glyph={trashIcon} />
-    </button>
-  );
 }
